inlined nodes seem to be stripped out smh
