+++ core/shared/src/main/scala-3/monocle/internal/focus/features/selectfield/SelectFieldParser.scala
@@ -30,18 +30,5 @@ private[focus] trait SelectFieldParser {
     getFieldType(fromType, fieldName, pos).flatMap { toType =>
       Right(FocusAction.SelectField(fieldName, fromType, getSuppliedTypeArgs(fromType), toType))
     }
-  // unappliedNamedTuple is the type lambda [Names, Values] =>> NamedTuple[Names, Values], used to harvest its type symbol later on
-  final class NamedTuples private (private val unappliedNamedTuple: Type[?]) {
-    def isNamedTuple(tpe: Type[?]) =
-      TypeRepr.of(using tpe).dealias.typeSymbol == TypeRepr.of(using unappliedNamedTuple).typeSymbol
-  }
 
-  object NamedTuples {
-    def create: Option[NamedTuples] =
-      Symbol
-        .requiredModule("scala.NamedTuple")
-        .declaredType("NamedTuple")
-        .headOption
-        .map(sym => NamedTuples(sym.typeRef.asType))
-  }
 }
